Fix after_filter deprecation warning on Rails 5
In Rails 4, the after_filter method was renamed to after_action. The
older method was still supported until Rails 5, which now prints a
deprecation warning when after_filter is used:

> DEPRECATION WARNING: after_filter is deprecated and will be removed in
> Rails 5.1. Use after_action instead.

With this change we'll use after_action if it's available, but will
continue to use after_filter on older versions of Rails.