Restore Config.api_key= with deprecation warning
We've removed the import rake task, along with the associated `api_key`
configuration option. This means that users who have added an API key to
their initializer will get this error when they update the gem:

    NoMethodError:  method `api_key=' for IntercomRails::Config:Class

This brings back the method, but without any behaviour; it just prints a
deprecation warning. This will allow those users to update to the latest
version of the gem without causing an error when their app boots.