Update README.mdown
It makes more sense to show this part of doc right after the developer generated the code. 
The first thing a dev does after generating the code is checking if the messenger shows up on his website. If it doesn't it is frustrating and the dev starts to inspect the console and the code when he just needs to turn on this config but didn't saw it in the docs.