Do not make assumptions about instance variables
This is a breaking change that removes what is a risky default in the
integration with the Rails app:

Assuming that `@user` is the currently logged in user, thus assuming we can
safely show potentielly personal messages to the currently logged in user.

After this change, you'll have to explicitly allow intercom-rails to dig into
your instance variables to find the currently logged in user, if you don't have
a `current_user` method.

This is done by configuring `config.user.current` with a Proc:

    config.user.current = Proc.new { @user }

or to preserve the previous behaviour:

    config.user.current = [
      Proc.new { current_user },
      Proc.new { @user }
    ]