Insert script immediately if loaded with Turbo (#352)
When the Messenger shim is loaded with Turbo, the load event has already
fired, so adding an event listener for it won't have any effect. Instead
we can immediately add the script tag to the page.

Also stop calling Intercom('reattach_activator'), which is an obsolete
API that no longer does anything.